+++ src/components/Footer.tsx
@@ -1,0 +1,17 @@
+export default function Footer() {
+  return (
+    <footer className="py-8 leading-relaxed text-xs">
+      <p>
+        Built with care by the{' '}
+        <a href="https://data.streetsforall.org">
+          Streets for All Data/Dev Team
+        </a>{' '}
+        with data from <a href="https://metro.net">LA Metro</a>
+      </p>
+      <p>
+        © 2025 <a href="https://streetsforall.org">Streets for All</a>
+      </p>
+      <p>🚌 🚲👩🏻‍🦽🚶🏾🌳</p>
+    </footer>
+  );
+}
